backend/requirements.txt
django==5.0.3
celery==5.3.6
django-celery-beat==2.6.0
gunicorn==21.2.0
kombu==5.3.5
psycopg2-binary==2.9.8
redis==5.0.2
djangorestframework==3.14.0
drf-spectacular==0.27.1
dj-rest-auth[with_social]==4.0.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
django-cors-headers==4.3.1
factory-boy==3.3.0
pillow==10.3.0
django-cloudinary-storage==0.3.0
urllib3==2.2.1
django-filter==24.2
stripe==9.6.0
gunicorn==21.2.0

black==24.2.0 ; extra == 'dev'
mypy==1.9.0 ; extra == 'dev'
pytest-cov==4.1.0 ; extra == 'dev'
pytest==8.1.1
pytest-django==4.8.0